retrofit2.0

+++ app/src/main/java/com/example/mariaa/weather/Main.java
@@ -7,33 +7,33 @@ public class Main {
 
     @SerializedName("temp")
     @Expose
-    private int temp;
+    private double temp;
     @SerializedName("pressure")
     @Expose
-    private double pressure;
+    private int pressure;
     @SerializedName("humidity")
     @Expose
     private int humidity;
     @SerializedName("temp_min")
     @Expose
-    private int tempMin;
+    private double tempMin;
     @SerializedName("temp_max")
     @Expose
-    private int tempMax;
+    private double tempMax;
 
-    public int getTemp() {
+    public double getTemp() {
         return temp;
     }
 
-    public void setTemp(int temp) {
+    public void setTemp(double temp) {
         this.temp = temp;
     }
 
-    public double getPressure() {
+    public int getPressure() {
         return pressure;
     }
 
-    public void setPressure(double pressure) {
+    public void setPressure(int pressure) {
         this.pressure = pressure;
     }
 
@@ -45,19 +45,19 @@ public class Main {
         this.humidity = humidity;
     }
 
-    public int getTempMin() {
+    public double getTempMin() {
         return tempMin;
     }
 
-    public void setTempMin(int tempMin) {
+    public void setTempMin(double tempMin) {
         this.tempMin = tempMin;
     }
 
-    public int getTempMax() {
+    public double getTempMax() {
         return tempMax;
     }
 
-    public void setTempMax(int tempMax) {
+    public void setTempMax(double tempMax) {
         this.tempMax = tempMax;
     }
